Add new controller for submit feedback
